refactor(sources): isolate portable pagination
