+++ crates/source-runtime-next/src/portable_ai.rs
@@ -4,6 +4,7 @@ mod adapter;
 mod catalog;
 mod langfuse;
 mod normalize;
+mod pagination;
 
 pub(crate) use adapter::PORTABLE_AI_SOURCE_EXECUTION_ADAPTERS;
